requirements.txt
google.generativeai
streamlit
pypdf
python-docx
pillow
opencv-python-headless
